+++ ntex/src/ws/client.rs
@@ -497,7 +497,7 @@ impl WsConnection<Sealed> {
                         description: None,
                     }))))
                 }
-                DispatchItem::Stop(Reason::KeepAliveTimeout) => Err(WsError::KeepAlive),
+                DispatchItem::Stop(Reason::KeepAlive) => Err(WsError::KeepAlive),
                 DispatchItem::Stop(Reason::ReadTimeout) => Err(WsError::ReadTimeout),
                 DispatchItem::Stop(Reason::WriteTimeout) => Err(WsError::WriteTimeout),
                 DispatchItem::Stop(Reason::Decoder(e)) => {